forensics/behind-enemy-lines/power_trip/src/requirements.txt
bcrypt==3.2.0
capstone==5.0.0rc2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colored-traceback==0.3.0
cryptography==36.0.2
cycler==0.11.0
ecdsa==0.17.0
fonttools==4.32.0
idna==3.3
intervaltree==3.1.0
kiwisolver==1.4.2
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
paramiko==2.10.3
Pillow==9.1.0
plumbum==1.7.2
psutil==5.9.0
pwntools==4.7.0
pycparser==2.21
pycryptodome==3.14.1
pyelftools==0.28
Pygments==2.11.2
PyNaCl==1.5.0
pyparsing==3.0.8
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.27.1
ROPGadget==6.7
rpyc==5.1.0
six==1.16.0
sortedcontainers==2.4.0
tlslite-ng==0.8.0a42
unicorn==2.0.0rc7
urllib3==1.26.9
